obs-webrtc: Generalize WebRTC Logic
Move logic that will be used by WHIP/WHEP into a helper file
