+++ plugins/obs-webrtc/whip-output.h
@@ -6,9 +6,9 @@
 #include <util/base.h>
 #include <util/dstr.h>
 
-#include <string>
 #include <atomic>
 #include <mutex>
+#include <string>
 #include <thread>
 
 #include <rtc/rtc.hpp>
